Rename `TokenName` module tree to `AssetName`.
In preparation for renaming the `TokenName` type to `AssetName`, this
commit makes the following module renamings:

```
- Cardano.Wallet.Primitive.Types.TokenName
+ Cardano.Wallet.Primitive.Types.AssetName
```
```
- Cardano.Wallet.Primitive.Types.TokenName.Gen
+ Cardano.Wallet.Primitive.Types.AssetName.Gen
```

Further commits will rename the `TokenName` type and related symbols.
