+++ lib/primitive/cardano-wallet-primitive.cabal
@@ -153,6 +153,8 @@ library
     Cardano.Wallet.Primitive.Types.AnyExplicitScripts
     Cardano.Wallet.Primitive.Types.AssetId
     Cardano.Wallet.Primitive.Types.AssetId.Gen
+    Cardano.Wallet.Primitive.Types.AssetName
+    Cardano.Wallet.Primitive.Types.AssetName.Gen
     Cardano.Wallet.Primitive.Types.Block
     Cardano.Wallet.Primitive.Types.Block.Gen
     Cardano.Wallet.Primitive.Types.BlockSummary
@@ -187,8 +189,6 @@ library
     Cardano.Wallet.Primitive.Types.TokenMap.Gen
     Cardano.Wallet.Primitive.Types.TokenMapWithScripts
     Cardano.Wallet.Primitive.Types.TokenMetadata
-    Cardano.Wallet.Primitive.Types.TokenName
-    Cardano.Wallet.Primitive.Types.TokenName.Gen
     Cardano.Wallet.Primitive.Types.TokenPolicyId
     Cardano.Wallet.Primitive.Types.TokenPolicyId.Gen
     Cardano.Wallet.Primitive.Types.TokenQuantity
