+++ lib/primitive/lib/Cardano/Wallet/Primitive/Types/TokenFingerprint.hs
@@ -14,11 +14,11 @@ module Cardano.Wallet.Primitive.Types.TokenFingerprint
 
 import Prelude
 
+import Cardano.Wallet.Primitive.Types.AssetName
+    ( TokenName (..)
+    )
 import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..)
-    )
-import Cardano.Wallet.Primitive.Types.TokenName
-    ( TokenName (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenPolicyId
     ( TokenPolicyId (UnsafeTokenPolicyId)
