+++ lib/primitive/test/spec/Cardano/Wallet/Primitive/Ledger/ConvertSpec.hs
@@ -19,6 +19,12 @@ import Cardano.Wallet.Primitive.Ledger.Convert
     , toLedgerTimelockScript
     , toWalletScript
     )
+import Cardano.Wallet.Primitive.Types.AssetName
+    ( TokenName
+    )
+import Cardano.Wallet.Primitive.Types.AssetName.Gen
+    ( genTokenNameLargeRange
+    )
 import Cardano.Wallet.Primitive.Types.Coin
     ( Coin (..)
     )
@@ -28,12 +34,6 @@ import Cardano.Wallet.Primitive.Types.TokenBundle
 import Cardano.Wallet.Primitive.Types.TokenBundle.Gen
     ( genTokenBundleSmallRange
     , shrinkTokenBundleSmallRange
-    )
-import Cardano.Wallet.Primitive.Types.TokenName
-    ( TokenName
-    )
-import Cardano.Wallet.Primitive.Types.TokenName.Gen
-    ( genTokenNameLargeRange
     )
 import Cardano.Wallet.Primitive.Types.TokenPolicyId
     ( TokenPolicyId
