+++ lib/wallet/src/Cardano/Wallet/TokenMetadata.hs
@@ -93,6 +93,9 @@ import Cardano.Wallet.Primitive.Types
 import Cardano.Wallet.Primitive.Types.AssetId
     ( AssetId (..)
     )
+import Cardano.Wallet.Primitive.Types.AssetName
+    ( TokenName (..)
+    )
 import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..)
     )
@@ -107,9 +110,6 @@ import Cardano.Wallet.Primitive.Types.TokenMetadata
     , validateMetadataName
     , validateMetadataTicker
     , validateMetadataURL
-    )
-import Cardano.Wallet.Primitive.Types.TokenName
-    ( TokenName (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenPolicyId
     ( TokenPolicyId (..)
